pyproject.toml
[tool.poetry]
name = "example-dishka-fastapiusers"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.112.2"}
dishka = "^1.3.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
async-sqlalchemy = "^1.0.0"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
src = ["src"]
exclude = ["src/example_dishka_fastapiusers/data_access/migrations/*"]
